message_scroll: Show scroll to bottom button.
Show/hide scroll to bottom button when the last message is
not visible in the current scroll position.

We adjust the bottom offset of the button based on compose box
height. We only do it for major height adjustments as there
is enough margin between compose and the button to take care
of minor height changes before it is hidden automatically.

Fixes #19862